services/evinfo-save-agent/requirements.txt
supabase
uuid
json
pika
python-dotenv